made some changes in file1

+++ file1.py
@@ -1,1 +1,1 @@
-print("int")
+print("float")
